+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This is intended as an overview of the major changes
 v0.4-SNAPSHOT
 ===
 - Fix parsing a device name with a space in it.
+- Parse the list of power failures
 
 v0.3
 ===